// src/generated/api/orders/models/adminOrderSummaryDtoFulfillmentStatus.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * DCTD-UTC Commerce API
 * Contract for storefront and admin applications
 * OpenAPI spec version: 1.0.0
 */

export type AdminOrderSummaryDtoFulfillmentStatus =
  (typeof AdminOrderSummaryDtoFulfillmentStatus)[keyof typeof AdminOrderSummaryDtoFulfillmentStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AdminOrderSummaryDtoFulfillmentStatus = {
  PENDING: 'PENDING',
  PICKING: 'PICKING',
  PACKED: 'PACKED',
  SHIPPED: 'SHIPPED',
  DELIVERED: 'DELIVERED',
  FAILED: 'FAILED',
  RETURNED: 'RETURNED',
  CANCELLED: 'CANCELLED',
} as const;