// src/generated/api/orders/models/adminOrderSummaryDtoPaymentStatus.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * DCTD-UTC Commerce API
 * Contract for storefront and admin applications
 * OpenAPI spec version: 1.0.0
 */

export type AdminOrderSummaryDtoPaymentStatus =
  (typeof AdminOrderSummaryDtoPaymentStatus)[keyof typeof AdminOrderSummaryDtoPaymentStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AdminOrderSummaryDtoPaymentStatus = {
  PENDING: 'PENDING',
  AWAITING_CONFIRMATION: 'AWAITING_CONFIRMATION',
  SUCCESS: 'SUCCESS',
  FAILED: 'FAILED',
  CANCELLED: 'CANCELLED',
  REFUNDED: 'REFUNDED',
} as const;