// src/generated/api/orders/models/adminOrderSummaryDtoStatus.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * DCTD-UTC Commerce API
 * Contract for storefront and admin applications
 * OpenAPI spec version: 1.0.0
 */

export type AdminOrderSummaryDtoStatus =
  (typeof AdminOrderSummaryDtoStatus)[keyof typeof AdminOrderSummaryDtoStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AdminOrderSummaryDtoStatus = {
  PENDING_CONFIRMATION: 'PENDING_CONFIRMATION',
  CONFIRMED: 'CONFIRMED',
  PICKING: 'PICKING',
  PACKED: 'PACKED',
  SHIPPED: 'SHIPPED',
  DELIVERED: 'DELIVERED',
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED',
} as const;