// src/generated/api/orders/models/adminOrderSummaryDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * DCTD-UTC Commerce API
 * Contract for storefront and admin applications
 * OpenAPI spec version: 1.0.0
 */
import type { AdminOrderSummaryDtoStatus } from './adminOrderSummaryDtoStatus';
import type { AdminOrderSummaryDtoPaymentStatus } from './adminOrderSummaryDtoPaymentStatus';
import type { AdminOrderSummaryDtoFulfillmentStatus } from './adminOrderSummaryDtoFulfillmentStatus';
import type { OrderRecipientDto } from './orderRecipientDto';

export interface AdminOrderSummaryDto {
  /** @pattern ^[1-9][0-9]*$ */
  id: string;
  orderNo: string;
  status: AdminOrderSummaryDtoStatus;
  paymentStatus: AdminOrderSummaryDtoPaymentStatus;
  fulfillmentStatus: AdminOrderSummaryDtoFulfillmentStatus;
  paymentMethod: string;
  shippingMethod: string;
  /** @pattern ^[1-9][0-9]*$ */
  branchId: string;
  branchName: string;
  warehouseName: string;
  grandTotal: string;
  itemCount: number;
  recipient: OrderRecipientDto;
  placedAt: string;
  version: number;
}